enable-vendor-revision.bbclass: Add it to append VENDOR_REVISION to PR
This bbclass is used for appending VENDOR_REVISION to PR from
vendor-revision.conf for the recipes which have patches.

The vendor-revision.conf is not a must, but without it, all the recipes which
patches will have a VR, there is no different in the first release such as
Yocto 5.1, but all the VRs will be updated from "vr51" to "vr511" in Yocto
5.1.1 since there is no history data (vendor-revision.conf).

Check the comments in the header of gen-vendor-revision.bbclass for more details.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>