autotools.bbclass: Copy gettext/po/Makefile.in.in to all po/Makefile.in.in
The previous code only copied to ${S}/po/Makefile.in.in, but there might be
other po/Makefile.in.in in ${S}, for example:
* bison:
  runtime-po/Makefile.in.in
  gnulib-po/Makefile.in.in

* gawk:
  extension/po/Makefile.in.in

* fontconfig
  po-conf/Makefile.in.in

* There might be more recipes in oe-core, and more recipes from other layers.

The build would be failed after upgrade gettext to 0.24.1 since gettext changed
the way to include Makevars, so we have to fix the recipes one by one, such as:
do_configure:prepend() {
    cp -f ${STAGING_DATADIR_NATIVE}/gettext/po/Makefile.in.in ${S}/runtime-po/
    cp -f ${STAGING_DATADIR_NATIVE}/gettext/po/Makefile.in.in ${S}/gnulib-po/
}

Even if we don't upgrade gettext to 0.24.1, have different po/Makefile.in.in in
the same ${S} is a kind of bug, this patch can fix the issues.

The fix way is:
Find all possible po/Makefile.in.in, if it is the same as ${S}/po/Makefile.in.in,
then override it with ${STAGING_DATADIR_NATIVE}/gettext/po/Makefile.in.in.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>